+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.Calendar/Classes/CalendarTimeSpan.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Krypton.Toolkit.Suite.Extended.Calendar
+{
+    public class CalendarTimeSpan
+    {
+        public DateTime Start { get; set; }
+        public DateTime End { get; set; }
+    }
+}
