+++ Source/Krypton Toolkit/Main/Krypton.Toolkit.Suite.Extended.Calendar/Events/CalendarItemEventArgs.cs
@@ -18,7 +18,7 @@ namespace Krypton.Toolkit.Suite.Extended.Calendar
         /// Creates a new <see cref="CalendarItemEventArgs"/>
         /// </summary>
         /// <param name="item">Related Item</param>
-        public CalendarItemEventArgs(CalendarItem item)
+        public CalendarItemEventArgs(CalendarItemAlternative item)
         {
             _item = item;
         }
@@ -27,12 +27,12 @@ namespace Krypton.Toolkit.Suite.Extended.Calendar
 
         #region Props
 
-        private CalendarItem _item;
+        private CalendarItemAlternative _item;
 
         /// <summary>
-        /// Gets the <see cref="CalendarItem"/> related to the event
+        /// Gets the <see cref="CalendarItemAlternative"/> related to the event
         /// </summary>
-        public CalendarItem Item
+        public CalendarItemAlternative Item
         {
             get { return _item; }
         }
